requirements.txt
flask>=3.0.0
pillow>=10.0.0
werkzeug>=3.0.0
requests==2.26.0
click==8.0.1
pytz
markdown
